let's replace that

+++ modular_skyrat/modules/soulstone_changes/components/return_on_death.dm
@@ -1,7 +1,7 @@
 //This component makes the ckey return to a body on death
 /datum/component/return_on_death
 	var/mob/sourcemob
-	var/soul_moved
+	var/deleting
 
 /datum/component/return_on_death/Initialize(mob/source)
 	if(!ismob(parent))
@@ -18,15 +18,16 @@
 		sourcemob.ckey = currentmob.ckey
 		if(HAS_TRAIT_FROM(sourcemob, TRAIT_SACRIFICED, "sacrificed"))
 			REMOVE_TRAIT(sourcemob, TRAIT_SACRIFICED, "sacrificed")
-		soul_moved = TRUE
 	else
 		to_chat(currentmob, span_warning("You were unable to return to your old body as it was destroyed."))
-	if(!QDELETED(src))
+	if(!QDELETED(src) && !deleting)
 		qdel(src)
+		deleting = TRUE
 
 
 /datum/component/return_on_death/Destroy(force, silent)
-	if(!soul_moved)
+	if(!deleting)
+		deleting = TRUE
 		return_to_old_body()
 	sourcemob = null
 	. = ..()
